+++ DataCore/Sql/Xml/XmlProductHelper.cs
@@ -368,7 +368,7 @@ public class XmlProductHelper
 		return description;
 	}
 
-	public ushort GetXmlShelfLifeDays(NomenclatureModel nomenclature, ushort shelfLifeDays)
+	public byte GetXmlShelfLifeDays(NomenclatureModel nomenclature, byte shelfLifeDays)
 	{
 		XmlProductModel xmlProduct = GetXmlProduct(nomenclature.Xml);
 		if (!xmlProduct.EqualsNew())
@@ -377,10 +377,10 @@ public class XmlProductHelper
 			string[] arr = strProductShelfLife.Split(' ');
 			if (arr.Length > 1)
 			{
-				if (ushort.TryParse(arr[0], out ushort productShelfLife2))
+				if (byte.TryParse(arr[0], out byte productShelfLife2))
 					return productShelfLife2;
 			}
-			if (ushort.TryParse(strProductShelfLife, out ushort productShelfLife1))
+			if (byte.TryParse(strProductShelfLife, out byte productShelfLife1))
 				return productShelfLife1;
 		}
 		return shelfLifeDays;
